WIP on map-link:
- update <map-extent> in lhin.mapml - add map-meta extent bounds.  
Default bounds of map-extent are those of the projection, which is now
expanding the bounds of the lhin layer defined by the map-meta in the
map-head.  Since the map-a in canada.mapml was linking to lhin.mapml
with target="_self" (replace self layer in map, zoom to bounds), it 
was zooming to the bounds of the projection, when the demo expects it
to zoom to the bounds of the lhin health unit area.