Add linking use case for map-link[rel=zoomin] and [zoomout] (#127)
* Add linking use case for map-link[rel=zoomin] and [zoomout], no
projection changing involved.

* map-link

- add styling/media-queries subdirectory, with example of selecting
preferred content for layer, based on mapml-extension, web-map-custom-element
map-link PRs.

* Add internationalization map-chrome-i18n folder and index.html

* map-link

- update internationalization example, add to index.html

* WIP on map-link

- update default-projection demo

Conflicts:
	change-projection/default-projection/index.html

* WIP on map-link:

- update cs attribute on map-geometry for demo

* - update vector tiles wgs84 bounds, zoom bounds to avoid errors

* WIP on map-link:

- update progressive enhancement example to eliminate console messages

* WIP on map-link:

- update bctile example to set native zoom range from 0 to 11

* map-link

- update BNG experiment to make a bit user friendlier

* WIP on map-link:

- update <map-extent> in lhin.mapml - add map-meta extent bounds.  
Default bounds of map-extent are those of the projection, which is now
expanding the bounds of the lhin layer defined by the map-meta in the
map-head.  Since the map-a in canada.mapml was linking to lhin.mapml
with target="_self" (replace self layer in map, zoom to bounds), it 
was zooming to the bounds of the projection, when the demo expects it
to zoom to the bounds of the lhin health unit area.

* [AUTO] Sync MapML Build

* Add linking use case for map-link[rel=zoomin] and [zoomout], no
projection changing involved.

* map-link

- add styling/media-queries subdirectory, with example of selecting
preferred content for layer, based on mapml-extension, web-map-custom-element
map-link PRs.

* Add internationalization map-chrome-i18n folder and index.html

* map-link

- update internationalization example, add to index.html

* WIP on map-link

- update default-projection demo

Conflicts:
	change-projection/default-projection/index.html

* WIP on map-link:

- update cs attribute on map-geometry for demo

* - update vector tiles wgs84 bounds, zoom bounds to avoid errors

* WIP on map-link:

- update progressive enhancement example to eliminate console messages

* WIP on map-link:

- update bctile example to set native zoom range from 0 to 11

* map-link

- update BNG experiment to make a bit user friendlier

* WIP on map-link:

- update <map-extent> in lhin.mapml - add map-meta extent bounds.  
Default bounds of map-extent are those of the projection, which is now
expanding the bounds of the lhin layer defined by the map-meta in the
map-head.  Since the map-a in canada.mapml was linking to lhin.mapml
with target="_self" (replace self layer in map, zoom to bounds), it 
was zooming to the bounds of the projection, when the demo expects it
to zoom to the bounds of the lhin health unit area.

---------

Co-authored-by: prushfor <prushfor@L-BSC-A146390.nrn.nrcan.gc.ca>